app/requirements.txt
pandas==2.2.2
pika~=1.3.2
numpy
requests~=2.32.3
pyyaml
opencv-python-headless==4.9.0.80
ultralytics
tqdm
prometheus-client==0.19.0